+++ lib/response.rs
@@ -77,10 +77,10 @@ impl Response {
     /// ```
     pub fn reason<T>(self, reason: T) -> Response
     where
-        T: fmt::Display,
+        T: AsRef<str>,
     {
         Response {
-            reason: Some(reason.to_string()),
+            reason: Some(reason.as_ref().to_owned()),
             ..self
         }
     }
@@ -99,10 +99,10 @@ impl Response {
     /// ```
     pub fn text<T>(self, text: T) -> Response
     where
-        T: fmt::Display,
+        T: AsRef<str>,
     {
         Response {
-            data: text.to_string().as_bytes().to_vec(),
+            data: text.as_ref().as_bytes().to_vec(),
             ..self
         }
     }